requirements.txt
streamlit>=1.31
openai-whisper>=20231117
faster-whisper>=1.1.0
pyttsx3>=2.98
